refactor: user 엔티티에 point 추가
