+++ module-api/src/main/java/com/example/api/facade/user/AddressFacade.java
@@ -21,6 +21,7 @@ public class AddressFacade implements CreateAddressUseCase {
 	private final AddressRepository addressRepository;
 	private final UserRepository userRepository;
 
+	@Override
 	@Transactional
 	public void createAddress(Long userId, AddressRequest.Create dto) {
